Fix node id collisions
According to @jseyfried, we actually don't need to be assigning node
ids to AST nodes in syntex. The problem we had was that we were
recursing into the macros looking for more macros, which ended up
assigning them a new node id, which conflicted with the invocation
node ids. If we simply just don't recurse, we don't have this problem.